refactor: reorganize imports and extract EmployeeTableProps interface into a separate types file for improved code structure and readability
